Don't use LinkDef files for ROOT 6
