--- FWCore/FWLite/src/FWCoreFWLiteLinkDef.h
@@ -1,6 +1,0 @@
-#include "FWCore/FWLite/interface/AutoLibraryLoader.h"
-#include "FWCore/FWLite/src/branchToClass.h"
-#ifdef __CINT__
-#pragma link C++ class AutoLibraryLoader;
-#pragma link C++ function branchToClass;
-#endif
