+++ FWCore/FWLite/src/classes.h
@@ -1,0 +1,2 @@
+#include "FWCore/FWLite/interface/AutoLibraryLoader.h"
+#include "FWCore/FWLite/src/branchToClass.h"
