+++ FWCore/FWLite/src/classes_def.xml
@@ -1,0 +1,5 @@
+<lcgdict>
+ <class name="AutoLibraryLoader" ClassVersion="10">
+   <version ClassVersion="10" checksum="981711899"/>
+ </class>
+</lcgdict>
