+++ IOPool/TFileAdaptor/src/TFileAdaptor.cc
@@ -256,25 +256,6 @@
     data.insert(std::make_pair("ROOT-tfile-write-totalMegabytes", w.str()));
   }
 
-/*
- * wrapper to bind TFileAdaptor to root, python etc
- * loading IOPoolTFileAdaptor library and instantiating
- * TFileAdaptorUI will make root to use StorageAdaptor for I/O instead
- * of its own plugins
- */
-class TFileAdaptorUI {
-public:
-
-  TFileAdaptorUI();
-  ~TFileAdaptorUI();
-
-  // print current Storage statistics on cout
-  void stats() const;
-
-private:
-  boost::shared_ptr<TFileAdaptor> me;
-};
-
 #include <iostream>
 
 TFileAdaptorUI::TFileAdaptorUI() {
