+++ IOPool/TFileAdaptor/src/classes.h
@@ -1,0 +1,3 @@
+#include "IOPool/TFileAdaptor/interface/TStorageFactoryFile.h"
+#include "IOPool/TFileAdaptor/interface/TStorageFactorySystem.h"
+#include "IOPool/TFileAdaptor/src/TFileAdaptor.h"
